backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
easyocr==1.7.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
deep-translator==1.11.4
google-generativeai==0.8.0
numpy==2.1.0
cachetools==5.4.0
